Add acceptance test for PuppetForge-based install of simp-simp_core
