+++ spec/acceptance/suites/forge_install/01_simp_server_bootstrap_spec.rb
@@ -1,0 +1,1 @@
+../default/01_simp_server_bootstrap_spec.rb
