+++ spec/acceptance/suites/forge_install/02_simp_server_rsync_spec.rb
@@ -1,0 +1,1 @@
+../default/02_simp_server_rsync_spec.rb
