+++ spec/acceptance/suites/forge_install/03_simp_server_ldap_spec.rb
@@ -1,0 +1,1 @@
+../default/03_simp_server_ldap_spec.rb
