+++ spec/acceptance/suites/forge_install/files
@@ -1,0 +1,1 @@
+../default/files
