+++ spec/acceptance/suites/forge_install/nodesets
@@ -1,0 +1,1 @@
+../default/nodesets/
